✨ feat(agent-runtime): host the request_human_approve executor in the package
Second Tier-A executor migration (LOBE-10949). request_human_approve now lives
in @lobechat/agent-runtime, driven by the StreamSink (event + chunk channels),
LifecycleSink (beforeHumanIntervention hook) and MessageTransport ports; the
server only registers the adapters via buildHost(). Behavior unchanged.

Contract refinements this migration forced:
- LifecycleSink.dispatch gains an opaque `serializedHooks` param — the
  per-operation webhook configs the executor reads off state.metadata._hooks.
- New ServerLifecycleSink adapter wraps HookDispatcher (binds operationId).
- ServerStreamSink.publishChunk now strips stepIndex from the chunk data
  (it is a positional arg to the stream manager, not part of the payload).

type-check clean; agent-runtime 266/266, RuntimeExecutors 125/125,
AgentRuntimeService 92/92.

Part of LOBE-10949
