+++ apps/server/src/modules/AgentRuntime/RuntimeExecutors.ts
@@ -2,6 +2,7 @@ import {
   type AgentInstruction,
   finish as createFinishExecutor,
   type InstructionExecutor,
+  requestHumanApprove as createRequestHumanApproveExecutor,
 } from '@lobechat/agent-runtime';
 
 import { buildHost } from './buildHost';
@@ -11,7 +12,6 @@ import { callTool } from './executors/callTool';
 import { callToolsBatch } from './executors/callToolsBatch';
 import { compressContext } from './executors/compressContext';
 import { execSubAgent, execSubAgents } from './executors/execSubAgent';
-import { requestHumanApprove } from './executors/humanApprove';
 import { resolveAbortedTools, resolveBlockedTools } from './executors/resolveTools';
 
 export { type RuntimeExecutorContext } from './context';
@@ -33,7 +33,7 @@ export const createRuntimeExecutors = (
     // Migrated into @lobechat/agent-runtime as part of the IO transport port
     // abstraction — the server now only registers adapters via buildHost.
     finish: createFinishExecutor(host),
-    request_human_approve: requestHumanApprove(ctx),
+    request_human_approve: createRequestHumanApproveExecutor(host),
     resolve_aborted_tools: resolveAbortedTools(ctx),
     resolve_blocked_tools: resolveBlockedTools(ctx),
   };
