+++ packages/agent-runtime/src/executors/index.ts
@@ -1,1 +1,2 @@
 export * from './finish';
+export * from './humanApprove';
