+++ packages/agent-runtime/src/transport/lifecycle.ts
@@ -17,9 +17,13 @@ export interface ToolCallMockResult {
  *
  * `dispatchBeforeToolCall` is the one interception point (returns a mock result
  * to skip real execution); everything else is fire-and-forget observation.
+ *
+ * `serializedHooks` carries the per-operation webhook configs (the server keeps
+ * them on `state.metadata._hooks` for production/queue mode); it is opaque to
+ * the package and forwarded verbatim to the adapter.
  */
 export interface LifecycleSink {
-  dispatch: (type: AgentHookType, event: AnyHookEvent) => Promise<void>;
+  dispatch: (type: AgentHookType, event: AnyHookEvent, serializedHooks?: unknown) => Promise<void>;
   dispatchBeforeToolCall: (
     event: Omit<ToolCallHookEvent, 'mock' | 'operationId'>,
   ) => Promise<ToolCallMockResult | null>;
